Add files via upload
Here I demonstated my understanding on how to use classes as a blueprint and to instantiate an object from that class to add basic information. I used the spring framework to help display the information on a localhost webpage. For the project, I used the following URL to display my webpage with the persons info: http://localhost:8080/ and http://localhost:8080/hello to show the page with regular text being displayed.